incidence editor fails if hitting apply for create incidences
Creating an event and hitting apply button make korganizer crash:
ASSERT: "item.isValid()" in file
kdepim/incidenceeditor-ng/incidencedialog.cpp,
line 487

The problem is that editoritemmanager do trigger a susses of the save
before updating the internal akonadi item. So wait for the itemFetchJob
to finish before emitting the success signal.